add CrudRepository

+++ src/main/java/io/kontur/controller/CatController.java
@@ -50,7 +50,7 @@ public class CatController {
       produces = MediaType.APPLICATION_JSON_VALUE)
   @ResponseStatus(HttpStatus.OK)
   public ResponseEntity<CatDto> findCatById(@PathVariable("id")
-                                            @Min(value = 1, message = "id must be positive") final long id) {
+                                            @Min(value = 1, message = "id must be positive") final int id) {
     log.info("get cat {}", id);
     CatDto dto = catService.read(id);
     linkModifier.withTagLocation(dto);
@@ -80,7 +80,7 @@ public class CatController {
       produces = MediaType.APPLICATION_JSON_VALUE)
   @ResponseStatus(HttpStatus.OK)
   public ResponseEntity<CatDto> deleteCatById(@PathVariable("id")
-                                              @Min(value = 1, message = "id must be positive") final long id) {
+                                              @Min(value = 1, message = "id must be positive") final int id) {
     log.info("get cat {}", id);
     catService.delete(id);
     return ResponseEntity.noContent().build();
